heavily refactored code, introduced consistent naming notation, relocated geocoder settings, and added method-level documentation
