+++ client/main.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Meteor } from 'meteor/meteor';
 import { render } from 'react-dom';
 
-import App from '../imports/ui/App.jsx';
+import App from '../imports/ui/app.jsx';
 
 Meteor.startup(() => {
   render(<App />, document.getElementById('render-target'));
