+++ imports/ui/my-map.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import GoogleMap from '../lib/GoogleMap';
+import GoogleMap from '../lib/google-map';
 
 export default class MyMap extends React.Component {
     constructor(props){
