+++ imports/ui/search-bar.jsx
@@ -2,8 +2,8 @@ import React, { Component, PropTypes } from 'react';
 import { createContainer } from 'meteor/react-meteor-data';
 import ReactDOM from 'react-dom';
 import { Times } from '../api/times.js';
-import Time from './Time.jsx';
-import MyMap from './MyMap.jsx';
+import Time from './time.jsx';
+import MyMap from './my-map.jsx';
 
 
 export default class SearchBar extends Component{
