+++ package.json
@@ -6,6 +6,7 @@
   },
   "dependencies": {
     "babel-runtime": "6.18.0",
+    "js-joda": "1.3.0",
     "meteor-node-stubs": "~0.2.0",
     "moment": "^2.17.1",
     "node-geocoder": "^3.16.0",
