+++ server/main.js
@@ -1,2 +1,18 @@
 import '../imports/api/times.js';
-import './database_initialization/DatabaseConstruction.js';
+import './database_initialization/database-construction.js';
+import '../imports/database/collections';
+import { RoomSlots, Rooms, Buildings } from '../imports/database/collections';
+import { persistRoomSlotsAsBuildings } from './database_initialization/database-construction'
+
+Meteor.methods({
+    /**
+     * Populates the Rooms collection with database scraped from ubc.ca. The scraped data is stored in roomSlots.json
+     */
+    'populateBuildingDatabase': function () {
+        RoomSlots.remove({});
+        Rooms.remove({});
+        Buildings.remove({});
+        const roomSlotsJSON = JSON.parse(Assets.getText('roomSlots.json'));
+        persistRoomSlotsAsBuildings(roomSlotsJSON);
+    }
+});
